# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Website Quản lý Lớp THCS REST API
  description: API phục vụ nền tảng số dành cho giáo viên chủ nhiệm, bộ môn, phụ huynh và học sinh THCS.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Production API Server v1
paths:
  /auth/login:
    post:
      summary: Đăng nhập tài khoản
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Đăng nhập thành công
        '401':
          description: Sai thông tin đăng nhập
        '429':
          description: Vượt quá số lần cho phép
...